+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project **DOES NOT** adhere to [Semantic Versioning](http://semver.org/spec
 ## [Unreleased]
 [Unreleased]: https://bitbucket.org/atlassian/db-replica/branches/compare/master%0Drelease-0.1.14
 
+### Fixed
+- NPE when calling `DualConnection#isReadOnly`
+
 ## [0.1.14] - 2020-12-07
 [0.1.14]: https://bitbucket.org/atlassian/db-replica/branches/compare/release-0.1.14%0Drelease-0.1.13